Fix deprecated null argument on ReflectionClass instantiation
When Sulu's AccessControlManager tries to get the permissions for a SecurityCondition which does not have an objectType, it calls the PhpcrAccessControlProvider::supports method with a null value. This is in turn passed as argument to the ReflectionClass contructor resulting in a deprecation.

This PR adds a check which prevents the invalid call.